pyproject.toml
[project]
name = "proxihud"
version = "0.1.0"
description = "AI Companion for Games"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "customtkinter>=5.2.2",
    "google-genai>=1.56.0",
    "keyboard>=0.13.5",
    "mss>=10.1.0",
    "packaging>=25.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
]

[project.scripts]
proxihud = "proxihud.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"